Move Document-types into `state/entities/contents`
